+++ modules/impex/core/src/sap/commerce/toolset/impex/lang/folding/ImpExMacroFoldingBuilder.java
@@ -139,7 +139,7 @@ public class ImpExMacroFoldingBuilder implements FoldingBuilder {
             }
         }
 
-        return resolvedValue;
+        return StringUtils.defaultIfBlank(resolvedValue, " ");
     }
 
     @NotNull